requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
pydantic==2.7.1
faiss-cpu==1.8.0.post1
chromadb==0.5.3
sentence-transformers==3.0.1
httpx==0.27.0
numpy==1.26.4
python-multipart==0.0.9
pdfplumber==0.11.0
rich==13.7.1

